Updating data source, data pre-filtering, and priors
I pulled in the scripts and raw data used to compile the data.csv file. I also added columns for the levels we may be interested in testing hypotheses for. There is some conforming of level naming that needs to be done and sorting out of the nesting structure that differs between data "owners" (e.g., Bohman and Giletto). For now, the analysis is still just run for Bohman data in Minnesota and Russet Burbank only.

I modified the priors a bit, but I don't think this is the issue. What I realized was that the problem wasn't with the model structure or the priors, but the fact that the underlying experimental data for some dates was not sufficient to fit a linear-plateau model and find the critical point. I came to this conclusion after considering for a long time how "clean" the data used in the Makowski et al. (2020) paper was and that we needed simple thresholds to filter the data and allow for only data that could plausibly fit a linear-plateau model. So I included a filter on the data prior to fitting the model where dates with a sd < 1.0 were excluded (in addition to the requirement that a date have at least 3 data points). The selection of 1.0 is arbitrary at best, and I also tried 0.5 as threshold value (with poorer success).

I used the brms model with date level group effects for Bmax and Si, and also removed two of the priors related to the sd for Bmax and Si which I didn't quite understand.

The predicted values for alpha_1 (4.71) and alpha_2 (0.45) are now well within the expected range. So I think the next steps are to nest additional random/group level effects for alpha1 and alpha2 to test the hypotheses of interest, right?